depend on pacman, add to base-devel (FS#62826)
pacman-key is run by this package in an install scriptlet, so we need to make
sure that pacman gets installed first by depending on it.

Since pacman will in turn not depend on archlinux-keyring any longer, we need
to add archlinux-keyring to base-devel so that it gets installed in clean
chroots that only preinstall base-devel.